IndexedDBへの処理をWorker以外でも行いたいので切り出す
